Allow ApiException to be suppressed
The new AllowAnyStatusCodeAttribute allows the ApiException, which is normally,
thrown on a response status code that does not indicate success, to be suppressed.

This is useful when the API consumer expects the API to return an 'invalid' status
code as part of normal operation, e.g. when a 404 indicates that a
resource does not exist, and the consumer is testing for existance of that resource.