Allow child interfaces, to some extent
Do not allow attributes on the child interface type itself. This can make
sense with a single layer of children (perhaps applying a header only to
methods inside that interface), but it gets tricky when a) children have
child interfaces (it's hard to determine what came from where,
essentially), and b) multiple interfaces define the same method.

However, we can grab all properties and methods.

Fixes #3