Get rid of PathParameterInfo / PathParameterInfo<T>
This value is going to be boxed anyway when we convert it to IFormattable,
which we will for all primitive types. Removing the genericness means
we can make PathParameterInfo a struct, and remove some allocation /
virtual method call overhead.

If we move to path parameter serializers, we can add the genericness back
in