Store generated types in the type system rather than a dictionary
This advantage of this approach is that we don't need a dictionary, and
we don't have to deal with 'object'. The downside is that we can't clear
the cache of creators. However, since the CLR prevents us from defining
the same type twice in one AppDomain *anyway*, having a cache that can't
be cleared makes a lot of sense.

Also make the ImplementationBuilder singleton: since it defines things like
the ModuleBuilder, I'm not actually happy with instantiating it twice.

Keep the same global lock which stops us from creating two types at the same
time. I considered putting the type creating in the cctor of the
TypeCreatorRegistry, which means the CLR prevents us from creating two
creators at the same time, without the need for explicit locking. However this
makes me a bit uncomfortable as it's not *entirely* predictable when cctors
will run, and it will also make confusing stack traces if there's an error.