Don't throw if a header could be applied to content, but there is no content
This is a nice little gotcha which is annoying to users. It was also impossible
to apply e.g. a Content-Type header to the whole class, then remove it on a
method, as we would fail to add the Content-Type header before we even got
near to removeing it.

Relates to #63