Make sure we always dispose HttpResponseMessage
HttpResponseMessage is only IDisposable so that the stream that
ReadAsStreamAsync returns is disposed, and only if
HttpCompletionOptions.ResponseHeadersRead is passed (otherwise it
uses a MemoryStream, which doesn't need disposing). So we were
never hitting a case where we wouldn't dispose something which
needed to be disposed. Regardless, we should be good and follow
the pattern. We can't dispose HttpResponseMessage if we return a
stream to the client, but in this case it's the client's responsibility
to dispose the Stream (which is all that the HttpResponseMessage
would have done).

Fixes #82