Allow content headers on a null [Body]
If they supplied a [Body] parameter, but they passed null, and they've put
content-related headers on the method, then add an empty content which
we can apply the headers to.

This addresses the specific case where someone needs to have e.g.
Content-Type on a body which is otherwise empty, and doesn't want to do this
by passing an empty body (but would rather pass null)

Fixes #109