Interpret Format as a string.Format placeholder if it looks like one
This lets people do more advanced things, such as 'merchantId{0}'.
There's a very small possibility of breakage here, if someone had a format
which looks very much like a string.Format format. However this is very
unlikely: none of the built-in format strings use { or } so they'd need
to be using a custom IFormattable object, and we add some other conditions
which string.Format formats must meet, but would be hard to do by chance.