Target netstandard2.1
It turns out that this includes System.Reflection.Emit, so we can lose our dependency
on that NuGet package

Add netcoreapp3.0 to the unit test targets to test this. This has an optimization which only
fetches NumberFormatInfo if actually needed when formatting a number, so make sure that we
use a format which requires this (i.e. currency, 'C')