Tweak how content headers behave again
If there's a content header on the interface, and the method
has [Body] (even if the user passed in null), then we still
add a dummy body with the content header.

This should align more to user expectations. If they've declared
the method with a [Body] parameter, then they're declaring that
that endpoint takes a body. If they pass null as the body, that's
saying that the body is empty, not that there is no body. Therefore
the content headers should still apply.

It also means that the headers sent for a particular method are
fixed at compile-time, rather than varying at runtime.